fix - #17: Bookmark 조회 핸들러메서드 이동
Related to: #17
